requirements.txt
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
pyodbc==5.0.1
loguru==0.7.2